requirements.txt
black==25.1.0
Django==5.2.4
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
drf-spectacular==0.28.0
markdown==3.8.2
tqdm==4.67.1
